# src/public/docs/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Earthid SSI Layer",
    "version": "2.0.0",
    "description": "This the SSI Layer of Earthid",
    "contact": {
      "email": "vishnuK@myearth.id",
      "name": "API Support"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Issuer Appnet - Dev"
    },
    {
      "url": "http://localhost:8081",
      "description": "Verifier Appnet - Dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "paths": {
    "/api/admin/getToken": {
      "get": {
        "summary": "Retrive Admin Token",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Access token  retrieved"
                    },
                    "data": {
                      "type": "object",
                      "description": "Admin User",
                      "properties": {
                        "token_type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "expires_in": {
                          "type": "string",
                          "example": "3599"
                        },
                        "ext_expires_in": {
                          "type": "string",
                          "example": "3599"
                        },
                        "access_token": {
                          "type": "string",
                          "description": "Id of credential",
                          "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsIng1dCI6IjlHbW55RlBraGMzaE91UjIybXZTdmduTG83WSIsImtpZCI6IjlHbW55RlBraGMzaE91UjIybXZTdmduTG83WSJ9.eyJhdWQiOiI2YThiNGIzOS1jMDIxLTQzN2MtYjA2MC01YTE0YTNmZDY1ZjMiLCJpc3MiOiJodHRwczovL3N0cy53aW5kb3dzLm5ldC9jOGRmYjY0MC04ODUyLTRlNTctYTYzYy1iYTlmNGM0ZjFjOGIvIiwiaWF0IjoxNjk5ODA0MDc4LCJuYmYiOjE2OTk4MDQwNzgsImV4cCI6MTY5OTgwNzk3OCwiYWlvIjoiRTJWZ1lIQUpPT241YzUrYnVrcndDbXZQVDNxMUFBPT0iLCJhcHBpZCI6ImU4NGQ2NzU4LWY0M2MtNDU1Yi1hZWE0LTkwZWIxNjczOGFhYSIsImFwcGlkYWNyIjoiMSIsImlkcCI6Imh0dHBzOi8vc3RzLndpbmRvd3MubmV0L2M4ZGZiNjQwLTg4NTItNGU1Ny1hNjNjLWJhOWY0YzRmMWM4Yi8iLCJvaWQiOiIxM2M2ZTIyNi1iMDQ0LTRlNWMtYTcwNi0zNmE0NjU0MmM0NmIiLCJyaCI6IjAuQVhFQVFMYmZ5RktJVjA2bVBMcWZURThjaXpsTGkyb2h3SHhEc0dCYUZLUDlaZk54QUFBLiIsInJvbGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsLkF1dGhvcml0eS5SZWFkIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ3JlZGVudGlhbC5SZXZva2UiLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5OZXR3b3JrLlJlYWQiLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5Db250cmFjdC5SZWFkIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ3JlZGVudGlhbC5TZWFyY2giLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5BdXRob3JpdHkuUmVhZFdyaXRlIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ29udHJhY3QuUmVhZFdyaXRlIl0sInN1YiI6IjEzYzZlMjI2LWIwNDQtNGU1Yy1hNzA2LTM2YTQ2NTQyYzQ2YiIsInRpZCI6ImM4ZGZiNjQwLTg4NTItNGU1Ny1hNjNjLWJhOWY0YzRmMWM4YiIsInV0aSI6IkNnY0I4OUQ0WTBtUXM5RFBULWRRQUEiLCJ2ZXIiOiIxLjAifQ.Kxu5PMteQ41PaXpViLIFv356YRWkXJCZej3Vf-ZSwkgcdYOEIM_lmT5yg22TCzMm1iWCWPwSlNvFeuF_Wqmyv12H1xFTWXwWJUSzFPLZ8sHOPsR0xn1IAevYqikKkNErjrMxx5cfwG-H1TLKGWsU0sFRwGuGc5Ve1P38fT0nUm-GVe2RonYrjB5fOP4qMF6Nch7ymcefn8GXFdOKbsHrVmpR29fb8c4rR9wkXaDlQcxGR5OEtzRi5lIWK3y6-VGVEHJ5mfZyqxUmhUyLYF1KXAwsoJUK8XK4zjzU4o91YkHgmuwBC2iZlUpcLfA5PTu_SwyFoIVO14elpFqEKU5xaw"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DID retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentia;"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/getAuyhorities": {
      "get": {
        "summary": "Get Authorities",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "List Authorities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Access token  retrieved"
                    },
                    "data": {
                      "type": "object",
                      "description": "Admin User",
                      "properties": {
                        "value": {
                          "type": "object",
                          "description": "Verifiable Credential of User",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "Id of credential",
                              "example": "MembershipType:be45c748-6e8c-4771-93a8-7529b033d0f3"
                            },
                            "type": {
                              "type": "string",
                              "description": "Type",
                              "example": [
                                "VerifiableCredential",
                                "MembershipType"
                              ]
                            }
                          },
                          "example": "Bearer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DID retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentia;"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/getContracts": {
      "get": {
        "summary": "Get contacts",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "authorizationId",
            "required": false,
            "schema": {
              "type": "string",
              "example": "e0c35664-2fc0-a48a-d30c-51c2eb77ab8d"
            },
            "description": "authorization id"
          },
          {
            "in": "query",
            "name": "contractId",
            "required": false,
            "schema": {
              "type": "string",
              "example": "99a27c9c-9213-68a6-bec5-0c4bab7bb633"
            },
            "description": "contract id"
          }
        ],
        "responses": {
          "200": {
            "description": "List Authorities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Access token  retrieved"
                    },
                    "data": {
                      "type": "object",
                      "description": "Admin User",
                      "properties": {
                        "token_type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "expires_in": {
                          "type": "string",
                          "example": "3599"
                        },
                        "ext_expires_in": {
                          "type": "string",
                          "example": "3599"
                        },
                        "access_token": {
                          "type": "string",
                          "description": "Id of credential",
                          "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsIng1dCI6IjlHbW55RlBraGMzaE91UjIybXZTdmduTG83WSIsImtpZCI6IjlHbW55RlBraGMzaE91UjIybXZTdmduTG83WSJ9.eyJhdWQiOiI2YThiNGIzOS1jMDIxLTQzN2MtYjA2MC01YTE0YTNmZDY1ZjMiLCJpc3MiOiJodHRwczovL3N0cy53aW5kb3dzLm5ldC9jOGRmYjY0MC04ODUyLTRlNTctYTYzYy1iYTlmNGM0ZjFjOGIvIiwiaWF0IjoxNjk5ODA0MDc4LCJuYmYiOjE2OTk4MDQwNzgsImV4cCI6MTY5OTgwNzk3OCwiYWlvIjoiRTJWZ1lIQUpPT241YzUrYnVrcndDbXZQVDNxMUFBPT0iLCJhcHBpZCI6ImU4NGQ2NzU4LWY0M2MtNDU1Yi1hZWE0LTkwZWIxNjczOGFhYSIsImFwcGlkYWNyIjoiMSIsImlkcCI6Imh0dHBzOi8vc3RzLndpbmRvd3MubmV0L2M4ZGZiNjQwLTg4NTItNGU1Ny1hNjNjLWJhOWY0YzRmMWM4Yi8iLCJvaWQiOiIxM2M2ZTIyNi1iMDQ0LTRlNWMtYTcwNi0zNmE0NjU0MmM0NmIiLCJyaCI6IjAuQVhFQVFMYmZ5RktJVjA2bVBMcWZURThjaXpsTGkyb2h3SHhEc0dCYUZLUDlaZk54QUFBLiIsInJvbGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsLkF1dGhvcml0eS5SZWFkIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ3JlZGVudGlhbC5SZXZva2UiLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5OZXR3b3JrLlJlYWQiLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5Db250cmFjdC5SZWFkIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ3JlZGVudGlhbC5TZWFyY2giLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5BdXRob3JpdHkuUmVhZFdyaXRlIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ29udHJhY3QuUmVhZFdyaXRlIl0sInN1YiI6IjEzYzZlMjI2LWIwNDQtNGU1Yy1hNzA2LTM2YTQ2NTQyYzQ2YiIsInRpZCI6ImM4ZGZiNjQwLTg4NTItNGU1Ny1hNjNjLWJhOWY0YzRmMWM4YiIsInV0aSI6IkNnY0I4OUQ0WTBtUXM5RFBULWRRQUEiLCJ2ZXIiOiIxLjAifQ.Kxu5PMteQ41PaXpViLIFv356YRWkXJCZej3Vf-ZSwkgcdYOEIM_lmT5yg22TCzMm1iWCWPwSlNvFeuF_Wqmyv12H1xFTWXwWJUSzFPLZ8sHOPsR0xn1IAevYqikKkNErjrMxx5cfwG-H1TLKGWsU0sFRwGuGc5Ve1P38fT0nUm-GVe2RonYrjB5fOP4qMF6Nch7ymcefn8GXFdOKbsHrVmpR29fb8c4rR9wkXaDlQcxGR5OEtzRi5lIWK3y6-VGVEHJ5mfZyqxUmhUyLYF1KXAwsoJUK8XK4zjzU4o91YkHgmuwBC2iZlUpcLfA5PTu_SwyFoIVO14elpFqEKU5xaw"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DID retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentia;"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/getDidDocument": {
      "get": {
        "summary": "Get did document",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "authorizationId",
            "required": false,
            "schema": {
              "type": "string",
              "example": "e0c35664-2fc0-a48a-d30c-51c2eb77ab8d"
            },
            "description": "authorization id"
          }
        ],
        "responses": {
          "200": {
            "description": "List Authorities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Access token  retrieved"
                    },
                    "data": {
                      "type": "object",
                      "description": "Admin User",
                      "properties": {
                        "token_type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "expires_in": {
                          "type": "string",
                          "example": "3599"
                        },
                        "ext_expires_in": {
                          "type": "string",
                          "example": "3599"
                        },
                        "access_token": {
                          "type": "string",
                          "description": "Id of credential",
                          "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsIng1dCI6IjlHbW55RlBraGMzaE91UjIybXZTdmduTG83WSIsImtpZCI6IjlHbW55RlBraGMzaE91UjIybXZTdmduTG83WSJ9.eyJhdWQiOiI2YThiNGIzOS1jMDIxLTQzN2MtYjA2MC01YTE0YTNmZDY1ZjMiLCJpc3MiOiJodHRwczovL3N0cy53aW5kb3dzLm5ldC9jOGRmYjY0MC04ODUyLTRlNTctYTYzYy1iYTlmNGM0ZjFjOGIvIiwiaWF0IjoxNjk5ODA0MDc4LCJuYmYiOjE2OTk4MDQwNzgsImV4cCI6MTY5OTgwNzk3OCwiYWlvIjoiRTJWZ1lIQUpPT241YzUrYnVrcndDbXZQVDNxMUFBPT0iLCJhcHBpZCI6ImU4NGQ2NzU4LWY0M2MtNDU1Yi1hZWE0LTkwZWIxNjczOGFhYSIsImFwcGlkYWNyIjoiMSIsImlkcCI6Imh0dHBzOi8vc3RzLndpbmRvd3MubmV0L2M4ZGZiNjQwLTg4NTItNGU1Ny1hNjNjLWJhOWY0YzRmMWM4Yi8iLCJvaWQiOiIxM2M2ZTIyNi1iMDQ0LTRlNWMtYTcwNi0zNmE0NjU0MmM0NmIiLCJyaCI6IjAuQVhFQVFMYmZ5RktJVjA2bVBMcWZURThjaXpsTGkyb2h3SHhEc0dCYUZLUDlaZk54QUFBLiIsInJvbGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsLkF1dGhvcml0eS5SZWFkIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ3JlZGVudGlhbC5SZXZva2UiLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5OZXR3b3JrLlJlYWQiLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5Db250cmFjdC5SZWFkIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ3JlZGVudGlhbC5TZWFyY2giLCJWZXJpZmlhYmxlQ3JlZGVudGlhbC5BdXRob3JpdHkuUmVhZFdyaXRlIiwiVmVyaWZpYWJsZUNyZWRlbnRpYWwuQ29udHJhY3QuUmVhZFdyaXRlIl0sInN1YiI6IjEzYzZlMjI2LWIwNDQtNGU1Yy1hNzA2LTM2YTQ2NTQyYzQ2YiIsInRpZCI6ImM4ZGZiNjQwLTg4NTItNGU1Ny1hNjNjLWJhOWY0YzRmMWM4YiIsInV0aSI6IkNnY0I4OUQ0WTBtUXM5RFBULWRRQUEiLCJ2ZXIiOiIxLjAifQ.Kxu5PMteQ41PaXpViLIFv356YRWkXJCZej3Vf-ZSwkgcdYOEIM_lmT5yg22TCzMm1iWCWPwSlNvFeuF_Wqmyv12H1xFTWXwWJUSzFPLZ8sHOPsR0xn1IAevYqikKkNErjrMxx5cfwG-H1TLKGWsU0sFRwGuGc5Ve1P38fT0nUm-GVe2RonYrjB5fOP4qMF6Nch7ymcefn8GXFdOKbsHrVmpR29fb8c4rR9wkXaDlQcxGR5OEtzRi5lIWK3y6-VGVEHJ5mfZyqxUmhUyLYF1KXAwsoJUK8XK4zjzU4o91YkHgmuwBC2iZlUpcLfA5PTu_SwyFoIVO14elpFqEKU5xaw"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DID retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentia;"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/onboard": {
      "get": {
        "summary": "Get onboard details",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "List details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Access token  retrieved"
                    },
                    "data": {
                      "type": "object",
                      "description": "Admin User",
                      "properties": {
                        "value": {
                          "type": "object",
                          "description": "Verifiable Credential of User",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "Id of credential",
                              "example": "MembershipType:be45c748-6e8c-4771-93a8-7529b033d0f3"
                            },
                            "type": {
                              "type": "string",
                              "description": "Type",
                              "example": [
                                "VerifiableCredential",
                                "MembershipType"
                              ]
                            }
                          },
                          "example": "Bearer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DID retrieval failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentia;"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errorCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}